Add files via upload
Working version of Movie class

+++ Classes/IndividualReview.java
@@ -1,24 +1,19 @@
 public class IndividualReview {
-    private int reviewRating;
+    private double reviewRating;
     private String reviewDescription;
-
-    public IndividualReview(int reviewRating, String reviewDescription) {
+    public IndividualReview(double reviewRating, String reviewDescription) {
         this.reviewRating = reviewRating;
         this.reviewDescription = reviewDescription;
     }
-
-    public int getReviewRating() {
+    public double getReviewRating() {
         return this.reviewRating;
     }
-
     public void setReviewRating(int reviewRating) {
         this.reviewRating = reviewRating;
     }
-
     public String getReviewDescription() {
         return this.reviewDescription;
     }
-    
     public void setReviewDescription(String reviewDescription) {
         this.reviewDescription = reviewDescription;
     }
